change : int->BigInteger, add Converter class (not finished yet)

+++ src/main/java/calculator/Calculator.java
@@ -4,6 +4,8 @@ import visitor.Countator;
 import visitor.Evaluator;
 import visitor.Stringator;
 
+import java.math.BigInteger;
+
 public class Calculator {
 
     /*
@@ -36,7 +38,7 @@ public class Calculator {
         printExpressionDetails(e, Notation.INFIX);
     }
 
-    public int eval(Expression e) {
+    public BigInteger eval(Expression e) {
         // create a new visitor to evaluate expressions
         Evaluator v = new Evaluator();
         // and ask the expression to accept this visitor to start the evaluation process
